+++ sample/src/main/java/io/snabble/testapp/CheckoutCustomerCardActivity.java
@@ -1,0 +1,14 @@
+package io.snabble.testapp;
+
+import androidx.fragment.app.Fragment;
+
+import io.snabble.sdk.ui.integration.CheckoutCustomerCardFragment;
+import io.snabble.sdk.ui.integration.CheckoutOfflineFragment;
+
+public class CheckoutCustomerCardActivity extends BaseActivity {
+    @Override
+    public Fragment onCreateFragment() {
+        return new CheckoutCustomerCardFragment();
+    }
+}
+